Added JWT signing and JWT verification.
The user will now get a Json Web Token (JWT) when authenticating. This token is the users ticket to the API, without this a user will need to re-log for a new token.

JWT verification has also been implemented so from now on devs will need to have a valid token when doing their testing through postman. The tokens are valid for 7 days though, so no need to renew all the time.

Furthermore, work has begun on the front-end, only the login form has been implemented thus far tho.