Code clean-up
Cleaned up the code that was commited earlier and removed a script that was no longer used (override.js).

The clean-up means that the query strings are no longer defined in the routes themselves, but rather in the database package. This means that the routes now only have the logic that they need. Along with this, the variable duplicates have been removed.

Also, the global variable (Database) has been abandoned in favor of the require('package') approach as it seems to be the prefered way to do it in the NodeJS community, something to do with testability and maintainability.