Registration + hashing
Registration has now been fully implemented and is available through the index.html, therefore the specific registration testing html file has been removed.

On top of that, hashing is also now in use on the client side of things. I have re-created the original 'testbruger' with the same password as before so that you can continue using that same account for your testing.

Also, Look-and-feel has been updated once again.