Database restructuring, security and package.json
The database has now been restructured so that fields are now in english rather than danish as we discussed. 

Also, a package.json file has been added to allow for people to get up and running faster.

And lastly, the 'secret' used for signing tokens which was previously just defined in the code has now been moved to a config file.