Authentication additions
Added a new method handler to the '/auth' route that handles 'get' calls and should be used for verifying the validity of tokens.

Also, this has been put to use in the client implementation. You will now be forwarded to the forum automatically if you have a valid token in your local storage already. Also, you will be redirected to the login page if you attempt to access the forum without a valid token.